+++ config/passport-local-strategy.js
@@ -1,5 +1,5 @@
 const LocalStrategy = require('passport-local').Strategy;
-const {User} = require('./database')
+const {User} = require('../database')
 const bcrypt = require('bcrypt');
 exports.initializingPassport = (passport)=>{
     passport.use (new LocalStrategy({
